requirements.txt
absl-py==0.2.2
astor==0.7.1
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.3
configparser==3.5.0
cycler==0.10.0
dask==0.18.1
decorator==4.3.0
defusedxml==0.5.0
dlib==19.15.0
easydict==1.9
entrypoints==0.3
enum34==1.1.6
feather-format==0.4.0
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
glog==0.3.1
graphviz==0.8.4
grpcio==1.13.0
h5py==2.9.0
html5lib==0.9999999
idna==2.6
imutils==0.4.6
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
kazoo==2.5.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.1.0
mat4py==0.4.2
matplotlib==2.2.2
mistune==0.8.4
mmdnn==0.2.3
mock==2.0.0
multiprocessing==2.6.2.1
mxnet==1.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
notebook==5.7.4
numpy==1.14.5
onnx==1.3.0
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.3
pbr==4.1.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.5.0
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
pyarrow==0.10.0
pycparser==2.19
Pygments==2.3.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-gflags==3.1.2
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.18.4
scandir==1.9.0
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.2.2
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
subprocess32==3.5.2
tb-nightly==1.10.0a20180716
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf2onnx==1.6.0
thrift==0.11.0
toolz==0.9.0
torch==0.1.12
tornado==5.1.1
tqdm==4.23.4
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.22
uuid==1.30
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.2
